srcディレクトリをルートディレクトリに変更
